Refactor: projection과 group by를 활영하여 사물함 상태 집계 최적화#27
