+++ src/main/java/org/univcabi/univcabi/cabinet/repository/CabinetCustomRepository.java
@@ -6,6 +6,8 @@ import org.univcabi.univcabi.cabinet.entity.BuildingName;
 import org.univcabi.univcabi.cabinet.entity.Cabinet;
 import org.univcabi.univcabi.cabinet.entity.CabinetHistory;
 import org.univcabi.univcabi.cabinet.entity.CabinetStatus;
+import org.univcabi.univcabi.cabinet.projection.CabinetStatusCountProjection;
+import org.univcabi.univcabi.cabinet.projection.CabinetStatusCountProjectionImpl;
 
 import java.util.List;
 import java.util.Optional;
@@ -28,4 +30,7 @@ public interface CabinetCustomRepository {
 
     // 사물함 상태로 사물함 정보 조회
     Page<Cabinet> findCabinetByStatus(CabinetStatus status, Pageable pageable);
+
+    // 빌딩별 사물함 상태 개수 정보 조회
+    List<CabinetStatusCountProjectionImpl> findCabinetStatusCountsGroupByBuilding();
 }
